media: one way into a room, for the two drivers that join one
The LiveKit driver held the whole of joining a room: the transport, the worker that runs it, the ear that turns arriving frames into audio, the token, and the delete that ends everything. A second driver is about to want every line of that and none of the SIP call underneath it, so the room half moves to a module of its own and the phone driver keeps only what is really the phone: the trunk, the dial, and the carrier statuses.

No behavior changes. The session also latches whether audio has flowed at all, which is what a driver waiting for somebody to turn up needs to know.
