+++ apps/simulator/src/egma_simulator/media/livekit.py
@@ -6,7 +6,8 @@ Two halves, both LiveKit's own:
   transport, purely outbound — signalling over a websocket it opens,
   media over ICE it negotiates — so the simulator needs no inbound
   network surface to conduct a phone call. Nothing dials the simulator;
-  the simulator dials.
+  the simulator dials. That half is :mod:`egma_simulator.media.room`,
+  shared with every other driver that reaches an agent through a room.
 - **The call.** LiveKit's SIP service places it, over a SIP trunk the
   deployment brings, and the answering phone appears in the room as an
   ordinary participant. Pipecat ships no example of this half, so the
@@ -39,118 +40,20 @@ from __future__ import annotations
 import asyncio
 import contextlib
 import logging
-import uuid
 
 from ..config import MediaSettings
 from ..redaction import SecretRegistry
-from ..speech import duration_seconds
 from . import ERROR, NOT_ANSWERED, MediaBackendError, sip_refusal
+from .room import (
+    QUOTED_REFUSAL_CHARS,
+    JoinedRoom,
+    RoomSession,
+    delete_room,
+    fresh_room_name,
+    room_token,
+)
 
 logger = logging.getLogger(__name__)
-
-ROOM_PREFIX = "egma-sim"
-"""What a simulation's room is called. One room per call, never reused —
-a room that outlived its call would put two simulations on one line."""
-
-PERSONA_IDENTITY = "egma-persona"
-"""Who the simulator is in the room. The *agent* under test is whoever
-answers the phone; this name is only ever the caller's."""
-
-CONNECT_SECONDS = 30.0
-"""How long joining the room may take before it counts as a bridge that
-cannot be reached."""
-
-TEARDOWN_SECONDS = 10.0
-"""How long a torn-down call may take to finish before it is cancelled."""
-
-QUOTED_REFUSAL_CHARS = 200
-"""How much of somebody else's refusal is quoted into a reason: enough to
-carry their own words about what was wrong, short of pasting a page."""
-
-
-async def _first_of(*events: asyncio.Event, within: float) -> bool:
-    """Wait until one of these happens, or say that none did."""
-    waiting = [asyncio.ensure_future(event.wait()) for event in events]
-    try:
-        done, _pending = await asyncio.wait(
-            waiting, return_when=asyncio.FIRST_COMPLETED, timeout=within
-        )
-    finally:
-        for unfinished in waiting:
-            if not unfinished.done():
-                unfinished.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await unfinished
-    return bool(done)
-
-
-class RoomSession:
-    """The audio of one call, once the room is joined.
-
-    Frames of the far end's audio arrive from the transport's input and
-    wait in a queue; the persona's audio goes out through the transport's
-    output, which paces it onto the wire the way a voice really travels.
-    """
-
-    def __init__(self, transport: object, *, band_hz: int) -> None:
-        self._transport = transport
-        self._band_hz = band_hz
-        self._heard: asyncio.Queue[bytes] = asyncio.Queue()
-        self._left = asyncio.Event()
-
-    @property
-    def sample_rate_hz(self) -> int:
-        return self._band_hz
-
-    @property
-    def far_end_left(self) -> bool:
-        return self._left.is_set()
-
-    def note_arrival(self, pcm: bytes) -> None:
-        self._heard.put_nowait(pcm)
-
-    def note_departure(self) -> None:
-        """The far end is off the line — the SIP participant left the room.
-
-        On a phone call there is no other signal and no better one: the
-        agent hanging up *is* the participant leaving.
-        """
-        self._left.set()
-
-    async def send(self, pcm: bytes) -> None:
-        """The persona's turn, said down the line, and waited out.
-
-        A voice takes as long to say a sentence as the sentence lasts, and
-        the transport writes the audio onto the wire at exactly that rate.
-        Returning before it is all said would start the far end's turn
-        while the persona was still talking — and every measurement of the
-        answer would carry the persona's own speaking time inside it.
-
-        What the line carried during all that is then dropped, because it
-        is the far end listening rather than the far end answering. An
-        agent that talks over the persona is lost with it: this seam
-        exchanges whole turns, so speech that overlaps two of them has
-        nowhere to go. What the record then shows is a conversation
-        without interruptions, which is true of what was measured and not
-        of what a real caller would have heard — worth knowing before
-        reading a transcript for barge-in behavior.
-        """
-        from pipecat.frames.frames import OutputAudioRawFrame
-
-        await self._transport.send_audio(
-            OutputAudioRawFrame(
-                audio=pcm, sample_rate=self._band_hz, num_channels=1
-            )
-        )
-        await asyncio.sleep(duration_seconds(pcm, self._band_hz))
-        while not self._heard.empty():
-            self._heard.get_nowait()
-
-    async def receive(self, seconds: float) -> bytes | None:
-        try:
-            return await asyncio.wait_for(self._heard.get(), timeout=seconds)
-        except TimeoutError:
-            return None
 
 
 class LiveKitBackend:
@@ -188,11 +91,8 @@ class LiveKitBackend:
         # second implementation of it.
         self._secrets = SecretRegistry()
         self._secrets.register(list(settings.secrets))
-        self._room_name = f"{ROOM_PREFIX}-{uuid.uuid4().hex}"
-        self._session: RoomSession | None = None
-        self._transport: object | None = None
-        self._running: asyncio.Task | None = None
-        self._worker: object | None = None
+        self._room_name = fresh_room_name()
+        self._room: JoinedRoom | None = None
         self._dialling: asyncio.Task | None = None
 
     @property
@@ -202,95 +102,18 @@ class LiveKitBackend:
 
     async def create_session(self) -> RoomSession:
         """Join the room, outbound, and answer with the call's audio."""
-        from pipecat.frames.frames import Frame, InputAudioRawFrame
-        from pipecat.pipeline.pipeline import Pipeline
-        from pipecat.pipeline.worker import PipelineParams, PipelineWorker
-        from pipecat.processors.frame_processor import FrameDirection, FrameProcessor
-        from pipecat.transports.livekit.transport import LiveKitParams, LiveKitTransport
-        from pipecat.workers.runner import WorkerRunner
-
-        transport = LiveKitTransport(
+        self._room = JoinedRoom(
             url=self._settings.livekit_url,
-            token=self._room_token(),
+            token=room_token(
+                self._settings.livekit_api_key,
+                self._settings.livekit_api_secret,
+                self._room_name,
+            ),
             room_name=self._room_name,
-            params=LiveKitParams(
-                audio_in_enabled=True,
-                audio_out_enabled=True,
-                audio_in_sample_rate=self._band_hz,
-                audio_out_sample_rate=self._band_hz,
-            ),
+            band_hz=self._band_hz,
+            quotable=self._quotable,
         )
-        self._transport = transport
-        session = RoomSession(transport, band_hz=self._band_hz)
-        joined = asyncio.Event()
-        refused = asyncio.Event()
-        told: list[str] = []
-
-        @transport.event_handler("on_connected")
-        async def _in_the_room(_transport: object) -> None:
-            joined.set()
-
-        @transport.event_handler("on_participant_disconnected")
-        async def _far_end_left(_transport: object, _participant: str) -> None:
-            session.note_departure()
-
-        class _Ear(FrameProcessor):
-            """Where the far end's audio leaves the transport and becomes
-            this call's audio."""
-
-            async def process_frame(
-                self, frame: Frame, direction: FrameDirection
-            ) -> None:
-                await super().process_frame(frame, direction)
-                if isinstance(frame, InputAudioRawFrame):
-                    session.note_arrival(frame.audio)
-                await self.push_frame(frame, direction)
-
-        worker = PipelineWorker(
-            Pipeline([transport.input(), _Ear(), transport.output()]),
-            params=PipelineParams(
-                audio_in_sample_rate=self._band_hz,
-                audio_out_sample_rate=self._band_hz,
-            ),
-            # The walk owns the clock and the limits; a transport that
-            # cancelled itself for being quiet would be a second, hidden
-            # limit with no record of having tripped.
-            idle_timeout_secs=None,
-            enable_turn_tracking=False,
-            enable_rtvi=False,
-        )
-
-        @worker.event_handler("on_pipeline_error")
-        async def _went_wrong(_worker: object, error: object) -> None:
-            # A transport that cannot reach the server says so *here* and
-            # nowhere else: the failure travels back up the pipeline as an
-            # error frame the library itself calls non-fatal, and the
-            # start frame reaches the end of the pipeline regardless. Read
-            # any other way, a room that was never joined would look like
-            # a room that was, and the call would be dialled into it.
-            told.append(str(getattr(error, "error", error)))
-            refused.set()
-
-        self._worker = worker
-        runner = WorkerRunner(handle_sigint=False)
-        await runner.add_workers(worker)
-        self._running = asyncio.create_task(runner.run(), name="phone-transport")
-
-        if not await _first_of(joined, refused, within=CONNECT_SECONDS):
-            raise MediaBackendError(
-                f"the livekit server at {self._settings.livekit_url} did not let the "
-                f"simulator into a room within {CONNECT_SECONDS:.0f}s",
-                ending=ERROR,
-            )
-        if refused.is_set():
-            raise MediaBackendError(
-                f"the livekit server at {self._settings.livekit_url} would not let "
-                f"the simulator into a room: {self._quotable('; '.join(told))}",
-                ending=ERROR,
-            )
-
-        self._session = session
-        return session
+        return await self._room.join()
 
     async def dial(self, number: str) -> None:
         """Ask LiveKit to place the call. Returns as soon as it is away."""
@@ -328,33 +151,25 @@ class LiveKitBackend:
                 dialling.cancel()
             with contextlib.suppress(asyncio.CancelledError, Exception):
                 await dialling
+        room, self._room = self._room, None
+        if room is None:
+            return
         try:
-            await self._end_transport()
+            await room.leave()
         finally:
-            await self._delete_room()
+            if room.joined:
+                # No room was ever joined, so there is no room to delete
+                # and nobody to ask — a trunk refused at construction must
+                # not cost a call to LiveKit that could only fail.
+                await delete_room(
+                    url=self._settings.livekit_url,
+                    api_key=self._settings.livekit_api_key,
+                    api_secret=self._settings.livekit_api_secret,
+                    room_name=self._room_name,
+                    quotable=self._quotable,
+                )
 
     # -- The parts that speak to LiveKit -------------------------------------
-
-    def _room_token(self) -> str:
-        """The persona's own way into its room, and nothing else's."""
-        from livekit import api
-
-        return (
-            api.AccessToken(
-                self._settings.livekit_api_key, self._settings.livekit_api_secret
-            )
-            .with_identity(PERSONA_IDENTITY)
-            .with_name(PERSONA_IDENTITY)
-            .with_grants(
-                api.VideoGrants(
-                    room_join=True,
-                    room=self._room_name,
-                    can_publish=True,
-                    can_subscribe=True,
-                )
-            )
-            .to_jwt()
-        )
 
     async def _place(self, number: str) -> str:
         """One `CreateSIPParticipant`, waited out, or the carrier's refusal.
@@ -420,55 +235,6 @@ class LiveKitBackend:
             with contextlib.suppress(Exception):
                 await lkapi.aclose()
         return participant.participant_identity
-
-    async def _end_transport(self) -> None:
-        if self._running is None:
-            return
-        from pipecat.frames.frames import EndFrame
-
-        try:
-            await self._worker.queue_frame(EndFrame())
-            await asyncio.wait_for(
-                asyncio.shield(self._running), timeout=TEARDOWN_SECONDS
-            )
-        except Exception as unfinished:
-            logger.warning("the call's transport did not end cleanly: %r", unfinished)
-        finally:
-            if not self._running.done():
-                self._running.cancel()
-                with contextlib.suppress(asyncio.CancelledError, Exception):
-                    await self._running
-            self._running = None
-
-    async def _delete_room(self) -> None:
-        if self._transport is None:
-            # No room was ever joined, so there is no room to delete and
-            # nobody to ask — a trunk refused at construction must cost a
-            # call to LiveKit that could only fail.
-            return
-        from livekit import api
-
-        lkapi = api.LiveKitAPI(
-            self._settings.livekit_url,
-            self._settings.livekit_api_key,
-            self._settings.livekit_api_secret,
-        )
-        try:
-            await lkapi.room.delete_room(
-                api.DeleteRoomRequest(room=self._room_name)
-            )
-        except Exception as unfinished:
-            # A room that was never created, or a server that cannot be
-            # reached to be told, has nothing left to be told. Teardown is
-            # not worth raising over — it would eat the walk's own answer.
-            logger.info(
-                "the room %s was not deleted: %s",
-                self._room_name,
-                self._quotable(repr(unfinished)),
-            )
-        finally:
-            with contextlib.suppress(Exception):
-                await lkapi.aclose()
 
     def _quotable(self, told: str) -> str:
         """Somebody else's words, minus this driver's secrets, short enough
